# src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "API MTIS - OpenAPI 3.0",
    "description" : "Practica grupal  de MTIS de la Universidad de Alicante 2022-2023.\n\n_Esta documentacion esta escrita en el Swagger Editor y con OpenAPI 3.0_\n\n`Autor: Nikita Polyanskiy, Serhii Vidernikov, Robert Kazaryan, Víctor Hernández Justamante y Ángel Tamame Pérez`",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "https://virtserver.swaggerhub.com/VHJ1_1/MTIS/1.0.0",
    "description" : "SwaggerHub API Auto Mocking"
  }, {
    "url" : "/"
  } ],
...
  "paths" : {
...
    "/Seguimiento/{id}" : {
      "get" : {
        "tags" : [ "Seguimiento" ],
        "summary" : "Obtener información de seguimiento por ID",
        "parameters" : [ {
          "name" : "RestKey",
          "in" : "header",
          "description" : "Api Key",
          "required" : false,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "id",
          "in" : "path",
          "description" : "ID del seguimiento a obtener",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Seguimiento"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "404" : {
            "description" : "Not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Seguimiento" ],
        "summary" : "Actualizar un seguimiento existente.",
        "operationId" : "updateSeguimiento",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "ID del seguimiento.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "$ref" : "#/components/requestBodies/Seguimiento"
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Seguimiento"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "404" : {
            "description" : "Not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Seguimiento" ],
        "summary" : "Abrir acceso a un seguimiento para el cliente.",
        "operationId" : "openAccesoSeguimiento",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "ID del seguimiento",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "404" : {
            "description" : "Not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Response"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Response" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "string",
            "description" : "Estado de la peticion",
            "example" : "Success, Bad request, etc."
          },
          "message" : {
            "type" : "string",
            "description" : "Mensaje mas concreto",
            "example" : "Se ha creado con exito, El campo no es valido, etc."
          }
        },
        "example" : {
          "message" : "Se ha creado con exito, El campo no es valido, etc.",
          "status" : "Success, Bad request, etc."
        }
      },
      "Seguimiento" : {
        "required" : [ "acceso", "estado", "identificador" ],
        "type" : "object",
        "properties" : {
          "identificador" : {
            "type" : "string",
            "description" : "Identificador del seguimiento.",
            "example" : "LN928756294"
          },
          "estado" : {
            "type" : "string",
            "description" : "Estado actual del seguimiento.",
            "example" : "Enviado al pais destinatario"
          },
          "acceso" : {
            "type" : "boolean",
            "description" : "Indica si el seguimiento esta disponible para el cliente"
          }
        },
        "example" : {
          "estado" : "Enviado al pais destinatario",
          "acceso" : true,
          "identificador" : "LN928756294"
        }
      },
...
    },
...
    "requestBodies" : {
...
      "Seguimiento" : {
        "description" : "Estructura para hacer peticiones",
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/Seguimiento"
            }
          }
        },
        "required" : true
      }
    }
  }
}